Add Win32 implementations of some unistd.h functions
We can't use the Win32 <io.h> header directly because Win32 doesn't define
ssize_t, and so we need to define that.  But the Windows implementations
of _read, _write, and possibly others assumes that ssize_t is 32-bits,
no matter what the size of size_t is.

We therefore define our own ssize_t as the equivalent size to size_t,
and also wrap _write and _read to only use SIZE_MAX bytes.