Make character tokens carry explicit character
This will simplify character handling significantly.

We use an explicit union to keep the size of the Token object down.

Currently on a 64-bit machine a Token is 128 + 32 + 32 = 196 bits (24
bytes).  But we should be able to reduce this to 128-bits by restricting
locations to 32-bits.  This would reduce a token to 16 bytes and so
reasonably able to be passed by value.