Fix (g)sub return values
We were assuming (g)sub returned the replaced string.  Whereas, they
modify the third parameter, and return the number of replacements.

This commit modifies the behaviour to do the right thing.

Unfortunately, this is the right thing in only some circumstances -
because we can match the empty string.  Need to rework the replacement
code.

Replacement formatting also allows back references to sub-matches which
is not valid.